modify memory numa headroom type from float64 to int64

+++ pkg/agent/qrm-plugins/memory/dynamicpolicy/state/checkpoint.go
@@ -28,7 +28,7 @@ var _ checkpointmanager.Checkpoint = &MemoryPluginCheckpoint{}
 type MemoryPluginCheckpoint struct {
 	PolicyName         string               `json:"policyName"`
 	MachineState       NUMANodeResourcesMap `json:"machineState"`
-	NUMAHeadroom       map[int]float64      `json:"numa_headroom"`
+	NUMAHeadroom       map[int]int64        `json:"numa_headroom"`
 	PodResourceEntries PodResourceEntries   `json:"pod_resource_entries"`
 	SocketTopology     map[int]string       `json:"socket_topology,omitempty"`
 	Checksum           checksum.Checksum    `json:"checksum"`
@@ -39,7 +39,7 @@ func NewMemoryPluginCheckpoint() *MemoryPluginCheckpoint {
 		PodResourceEntries: make(PodResourceEntries),
 		MachineState:       make(NUMANodeResourcesMap),
 		SocketTopology:     make(map[int]string),
-		NUMAHeadroom:       make(map[int]float64),
+		NUMAHeadroom:       make(map[int]int64),
 	}
 }
